Add files via upload
Majority of the students with mental health issues do not speak up about their problems due to fear of judgement and stigmatization. They bottle their problems until it is too much and as a result, they end up making irreversible decisions. Stigma, particularly surrounding mental health disorders and suicide means many people thinking of taking their own life or who have attempted suicide are not seeking help and are therefore not getting the help they need. Denial of the problem and the silence that surrounds mental health issues only serve to make things worse. All this is because these young people do not have a safe space to access the treatment that they need and speak about their problems freely.
With the burden of increasing mental health problems, the current health care systems in place are either not readily available or not confidential enough for youths to feel safe and open up their problems.(Vidourek et al., 2014)
Siritalk application  provides a safe, confidential space for the mentally ill.